feat: update

+++ dev_ifcu_remote_broker/Globals.h
@@ -5,14 +5,14 @@
 #define REQUEST_VALUES_LEN 4
 #define RESPONSE_VALUES_LEN 7
 
-int requestValues[REQUEST_VALUES_LEN] = {
+uint16_t requestValues[REQUEST_VALUES_LEN] = {
   0,  // on/off
   0,  // set point temp
   0,  // mode
   0,  // speed
 };
 
-int responseValues[RESPONSE_VALUES_LEN] = {
+uint16_t responseValues[RESPONSE_VALUES_LEN] = {
   0,  // on/off
   0,  // room temp
   0,  // set point temp
